requirements.txt
dvc
scikit-learn
pandas
numpy
joblib
mlflow
watchdog
pyyaml
gradio 
